requirements.txt
# LLM & AI
openai>=1.0.0
anthropic>=0.18.0

# Trading
coinbase-advanced-py>=1.8.0
requests>=2.31.0
cryptography>=42.0.0
dydx-v4-client>=1.0.0

# Social Media
tweepy>=4.14.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0